requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-magic==0.4.27
APScheduler==3.10.4
